+++ core/src/main/java/io/github/arkosammy12/jemu/core/cosmacvip/CosmacVIPEmulator.java
@@ -133,7 +133,7 @@ public class CosmacVIPEmulator implements CDP1802System, CDP1802.SystemBus, Rese
     }
 
     @Override
-    public int getFramerate() {
+    public double getFramerate() {
         return this.frameRate;
     }
